+++ lib/controllers/change-password.js
@@ -67,7 +67,7 @@ module.exports = function (req, res, next) {
       var formActionUri = config.web.changePassword.uri
         + '?sptoken=' + encodeURIComponent(req.query.sptoken);
 
-      return helpers.handleMultiTenantRequests(req, res, formActionUri, function () {
+      return helpers.resolveOrganization(req, res, formActionUri, function () {
         application.verifyPasswordResetToken(sptoken, function (err, result) {
           if (err) {
             logger.info('A user attempted to reset their password with a token, but that token verification failed.');
