+++ lib/controllers/forgot-password.js
@@ -48,7 +48,7 @@ module.exports = function (req, res, next) {
 
       var formActionUri = config.web.forgotPassword.uri;
 
-      return helpers.handleMultiTenantRequests(req, res, formActionUri, function () {
+      return helpers.resolveOrganization(req, res, formActionUri, function () {
         forms.forgotPasswordForm.handle(req, {
           // If we get here, it means the user is submitting a password reset
           // request, so we should attempt to send the user a password reset email.
