+++ lib/controllers/login.js
@@ -112,7 +112,7 @@ module.exports = function (req, res, next) {
           helpers.render(req, res, view, options);
         }
 
-        return helpers.handleMultiTenantRequests(req, res, formActionUri, function () {
+        return helpers.resolveOrganization(req, res, formActionUri, function () {
           helpers.setTempCookie(res, 'oauthRedirectUri', req.originalUrl);
 
           forms.loginForm.handle(req, {
