+++ lib/controllers/register.js
@@ -203,7 +203,7 @@ module.exports = function (req, res, next) {
         var writeFormError = helpers.writeFormError.bind(null, req, res, view, viewModel);
         var formActionUri = config.web.register.uri;
 
-        return helpers.handleMultiTenantRequests(req, res, formActionUri, function () {
+        return helpers.resolveOrganization(req, res, formActionUri, function () {
           switch (req.method) {
             // We should render the registration template.
             case 'GET':
