+++ lib/controllers/verify-email.js
@@ -81,7 +81,7 @@ module.exports = function (req, res, next) {
         formActionUri += '?sptoken=' + encodeURIComponent(req.query.sptoken);
       }
 
-      return helpers.handleMultiTenantRequests(req, res, formActionUri, function () {
+      return helpers.resolveOrganization(req, res, formActionUri, function () {
         forms.resendAccountVerificationEmailForm.handle(req, {
           // If we get here, it means the user is submitting a request to resend their
           // account verification email, so we should attempt to send the user another
