+++ lib/helpers/index.js
@@ -29,5 +29,5 @@ module.exports = {
   parentDomainRedirect: require('./parent-domain-redirect'),
   isRootDomainMultiTenantRequest: require('./is-root-domain-multi-tenant-request'),
   redirectToOrganization: require('./redirect-to-organization'),
-  handleMultiTenantRequests: require('./handle-multi-tenant-requests')
+  resolveOrganization: require('./resolve-organization')
 };
