+++ lib/helpers/resolve-organization.js
